avcodec/vvcdec: fix undefined behaviours for derive_affine_mvc
libavcodec/vvc/vvc_inter.c:823:18: runtime error: signed integer overflow: 1426128896 + 1426128896 cannot be represented in type 'int'
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior libavcodec/vvc/vvc_inter.c:823:18

Suggested-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>

+++ libavcodec/vvc/vvc_inter.c
@@ -817,10 +817,13 @@ static void derive_affine_mvc(MvField *mvc, const VVCFrameContext *fc, const MvF
     const int vs = fc->ps.sps->vshift[1];
     const MvField* mv2 = ff_vvc_get_mvf(fc, x0 + hs * sbw, y0 + vs * sbh);
     *mvc = *mv;
-    mvc->mv[0].x += mv2->mv[0].x;
-    mvc->mv[0].y += mv2->mv[0].y;
-    mvc->mv[1].x += mv2->mv[1].x;
-    mvc->mv[1].y += mv2->mv[1].y;
+
+    // Due to different pred_flag, one of the motion vectors may have an invalid value.
+    // Cast them to an unsigned type to avoid undefined behavior.
+    mvc->mv[0].x += (unsigned int)mv2->mv[0].x;
+    mvc->mv[0].y += (unsigned int)mv2->mv[0].y;
+    mvc->mv[1].x += (unsigned int)mv2->mv[1].x;
+    mvc->mv[1].y += (unsigned int)mv2->mv[1].y;
     ff_vvc_round_mv(mvc->mv + 0, 0, 1);
     ff_vvc_round_mv(mvc->mv + 1, 0, 1);
 }
